Change casts of function pointers from int to long. This makes the compiler warnings go away, but the compiler is throwing away 32-bits as the long value is silently truncated to an int on alpha. But the program works, so that must not matter.

+++ games/trek/phaser.c
@@ -138,7 +138,7 @@ phaser()
 	if (!manual)
 	{
 		ptr = getcodpar("Manual or automatic", Matab);
-		manual = (int) ptr->value;
+		manual = (long) ptr->value;
 	}
 	if (!manual && damaged(COMPUTER))
 	{
